pyproject.toml
[tool.poetry]
name = "oz-project-03-be"
version = "0.1.0"
description = ""
authors = ["skwwnl <skwwnl@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.7"
fastapi = {version = "0.115.6", extras = ["standard"]}
uvicorn = "0.32.0"
alembic = "1.14.0"
sqlalchemy = {extras = ["asyncio"], version = "2.0.36"}
asyncpg = "0.30.0"
python-dotenv = "1.0.1"
redis = {extras = ["async"], version = "5.2.0"}
ulid = "1.1"
pyjwt = "2.9.0"
email-validator = "2.2.0"
argon2-cffi = "23.1.0"
python-jose = "3.3.0"
requests = "2.32.3"
aiosmtplib = "3.0.2"
boto3 = "1.35.69"
python-multipart = "0.0.17"
odmantic = "1.0.2"
motor = {extras = ["srv"], version = "3.6.0"}
openai = "1.55.3"
aiokafka = "0.12.0"
websockets = "14.1"
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
coverage = "^7.6.4"
pytest = "^8.3.3"
psycopg2-binary = "^2.9.10"
pytest-asyncio = "^0.24.0"
types-python-jose = "^3.3.4.20240106"

[tool.black]
line-length = 150

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"

